+++ auth/auth-service/src/main/java/shop/woowasap/auth/service/dto/request/UserCreateRequest.java
@@ -1,7 +1,6 @@
 package shop.woowasap.auth.service.dto.request;
 
 import lombok.Builder;
-import shop.woowasap.auth.domain.User;
 
 @Builder
 public record UserCreateRequest(
@@ -11,7 +10,4 @@ public record UserCreateRequest(
 
 ) {
 
-    public User toDomain() {
-        return User.builder().userId(userId).password(password).build();
-    }
 }
